modified add pet form content, added basic firebase calls on form submission to add pet to database.  Separated out the add pet form from the add pet screen. Changed navigation from admin_homepage "add pet" button to the add pet screen with the add pet form.
