+++ plentyofpets/lib/components/admin_homepage_buttons.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import '../screens/add_pet_screen.dart';
+//import '../screens/add_pet_screen.dart';
+import '../screens/add_edit_pet_screen.dart';
 
 
 class AddPetCardButton extends StatelessWidget {
