--- plentyofpets/lib/screens/add_pet_screen.dart
@@ -1,22 +1,0 @@
-import 'package:flutter/material.dart';
-
-class AddPetScreen extends StatelessWidget {
-  const AddPetScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Second Route'),
-      ),
-      body: Center(
-        child: ElevatedButton(
-          onPressed: () {
-            Navigator.pop(context);
-          },
-          child: const Text('Go back!'),
-        ),
-      ),
-    );
-  }
-}
