+++ plentyofpets/pubspec.yaml
@@ -40,6 +40,8 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  form_builder_validators: ^7.4.0
+  form_builder_image_picker: ^2.0.0
 
 dev_dependencies:
   flutter_test:
